Move GetPartial/SetPartial (#2345)
Summary:
This is part of a larger set of features for the object model to fully model the Set/Get phases for partial and abstract objects. I wanted to break it down into smaller steps.

This is just a plain move of GetPartial and SetPartial to get.js/properties.js. All other property modeling for concrete keys are already there. You can tell that a lot of imports and logic becomes duplicated because they do similar things.

An alternative could be to move all the logic into ObjectValue but that's not how the spec is structured.
This also causes trouble for me trying to generalize these operations for abstract object values which may not consult the ObjectValue's virtual dispatch.
Pull Request resolved: https://github.com/facebook/prepack/pull/2345

Differential Revision: D9105679

Pulled By: sebmarkbage

fbshipit-source-id: 9d77d633eb238e4dd46e2ad43485fc885e85152c