Fix issue 2266 (#2303)
Summary:
Release Notes: None

Fixes #2266.

Looking at the `__optimize` code, what we want is for every call to `__optimize` to optimize that function, even in conditonal contexts. In this case, it makes no sense to store the functions to optimize in a special global because we don't care about the applying/reverting of effects. It simplifies the code significantly to be able to store these values in the realm instead.

The only case where this would matter is if effects containing an __optimize call get created then discarded (I am unsure if there are any legitimate cases of this in Prepack at the moment, but I created a CompilerDiagnostic just in case).
Pull Request resolved: https://github.com/facebook/prepack/pull/2303

Differential Revision: D9134968

Pulled By: cblappert

fbshipit-source-id: 040f7ccc24f928e6b8daa068521d3848caffc4d2