Remove makeNotPartial and pass around explicit boolean instead (#2357)
Summary:
Release notes: none

This PR removes the `makeNotPartial` method and all its call-sites in favour of explicitly telling execution paths that we want own properties even if partial. This stops the mutation of objects where we set the partial boolean flag, which actually is a very common cause of side-effects in pure functions. Furthermore, the whole `makeNotPartial` before and after was potentially error-prone if the `try/catch` was omitted by mistake and it was generally a code smell.

I also made it test-runner log the failing test name upon error, which helps with debugging.
Pull Request resolved: https://github.com/facebook/prepack/pull/2357

Differential Revision: D9137606

Pulled By: trueadm

fbshipit-source-id: c0c59615f7a7d46836d26a3b2aecfb89274933d8