Fix issue #2359 (#2382)
Summary:
Release Notes: None

We weren't properly adding leaked bindings to the optimized function's FunctionInstance, so state didn't get cleaned up properly between passes of the serializer.

There was also a bug in emitter invariants. Loosening the invariant causes the test cases to pass.

Also the invariant seems to have been incorrect in the first place. I suspect it is just the direct negation of [this line](https://github.com/facebook/prepack/blob/393624f00f7ad5291622ee28bf90ea0470becbcf/src/serializer/Emitter.js#L103).

Added slightly minimized repros from the issue.

I am not quite sure what we are trying to test with the last line of this invariant. I think what we're trying to test with the last line of this invariant is: If we're emitting to an additional function, it must be the case that referencedDeclaredValues has `val`. Here the test also passes when `referencedDeclaredValues.has(val)` is `false` which seems wrong to me.

I'm also dubious that top level optimized functions don't consider the MainGenerator as their parent generator, as this means that any derived ids declared in the MainGenerator shouldn't be considered declared in optimized functions (was able to repro that in conditions2 test after tightening invariants). For the sake of unblocking React Compiler, I'll look into that in a future PR.

Solves #2359.
Pull Request resolved: https://github.com/facebook/prepack/pull/2382

Differential Revision: D9158973

Pulled By: cblappert

fbshipit-source-id: a6915d0f125708739022d8576d5f007107e76b6b