Improve defaultProps handling (#2391)
Summary:
Release notes: none

Fixes a `defaultProps` bug and also improves bloat slightly by ensuring we state that the default props helper can be omitted if the value it mutates is never used. Furthermore, added `isPure` to the binary expression temporal, so if the value created from it never gets used, then we don't emit the action.
Pull Request resolved: https://github.com/facebook/prepack/pull/2391

Differential Revision: D9179550

Pulled By: trueadm

fbshipit-source-id: b5671a725e347be3876a04b927eef99977757bd8