Fix default attributes in joined descriptor (#2390)
Summary:
When we create a joined descriptor in an abstract object we need to preserve any attributes of that property that were already there. To do that, we just need to read the first descriptor of the elements since all elements must have the same attributes (or we fatal).
Pull Request resolved: https://github.com/facebook/prepack/pull/2390

Differential Revision: D9193129

Pulled By: sebmarkbage

fbshipit-source-id: 32db3e88ae65de9a09ff6de72ed5b91ec02cf090