Fix tryGetOptimizedFunctionRoot to actually get optimized function root (#2401)
Summary:
Release Notes: None

Previously, `tryGetOptimizedFunctionRoot` would only get the root successfully if the root was one of the optimized functions or the parent of one of the residual functions. Now we look at the optimized functions as well as any of their parents. Added 2 test cases.

Fixes #2399
Pull Request resolved: https://github.com/facebook/prepack/pull/2401

Differential Revision: D9242153

Pulled By: cblappert

fbshipit-source-id: 9bdffce23d5dd19d9ce5a6b09e1ee9647ca6e5cb