Remove React pathCondition resetting (#2415)
Summary:
Release notes: none

During React reconciliation we wrongly reset `pathConditions`. This was done originally because of a pushing false invariant, that I incorrectly thought was a bug because of React specific code.

However, today, I re-worked the original test code to remove all React logic and I found that the invariant occurs even without React specific logic. This strongly suggests that this is still an existing bug in Prepack and adding `pathConditions` logic, which gets around it, was the wrong strategy and should be reverted.
Pull Request resolved: https://github.com/facebook/prepack/pull/2415

Differential Revision: D9264730

Pulled By: trueadm

fbshipit-source-id: 2d3703e638c894eb0929b9a30a7d74ccbd998ea3