Clarify terminology for Leaked Value Analysis (#2444)
Summary:
This PR continues the break up of #2185 into smaller pieces, and makes the use of Leaking, Havocing and Materialization consistent with agreed upon terminology.

*Leaking* transitively adds mutable locations to the environment. Accessing such leaked locations is equivalent to accessing to the environment.

*Havocing* takes on its usual meaning in static analysis: to set an abstract value to "unknown" i.e. topVal.

*Materialization* snapshots locations into runtime assignments for the purpose of outlining functions.

Presently, leaking always leads to havocing, and materialization. But in addition, it also marks a location or an object as "leaked" implying that it may be aliased and/or no longer be "well behaved" (for example, a leaked object may leak any object passed to one of its methods).
Pull Request resolved: https://github.com/facebook/prepack/pull/2444

Differential Revision: D9392730

Pulled By: sb98052

fbshipit-source-id: 0ef987be8a71bb74e2def0526caf203bbfc34d82