Simplify widened array index access test (#2453)
Summary:
This test bedeviled me for quite some time because I thought it was testing for aliasing effects - it isn't. This is my fault, since I posted this in my repro. Aliasing effects are still to be dealt with in nested optimized functions. Anyhow, simplifying to prevent somebody else from having this confusion.
Pull Request resolved: https://github.com/facebook/prepack/pull/2453

Reviewed By: trueadm

Differential Revision: D9420733

Pulled By: sb98052

fbshipit-source-id: 755ccd90b1e64afdfe6ec627a3306e8baf5952ea