Remove numeric widened property special case in abstract get
I'm refactoring these parts in a different diff and stumble upon this.

I don't really get (pun intended) why this is here. It is also returning
a property binding instead of a descriptor which doesn't make sense.

Removing it doesn't fail any tests so at the very least we need to add
coverage to this.