Don't complain about infeasible paths unless there has been a k-limit (#2472)
Summary:
Release note: Make simplification more robust when complexity limits are reached

Fixes #2361

Any particular simplification can fail because Prepack has run out of a global limit for allowable implications (set by the abstractValueImpliesMax option). If Prepack then recovers from from the exception and carries on, it may later do the same simplification and succeed. If this happens in a context where the now known result is not expected, there will be an invariant failure, as for example in the issue above.

The invariant useful for finding bugs when reasoning is precise, so I'm keeping it, but only if no limit was exceeded.
Pull Request resolved: https://github.com/facebook/prepack/pull/2472

Reviewed By: trueadm

Differential Revision: D9479598

Pulled By: hermanventer

fbshipit-source-id: c38c49257df1c1e207e01a24b92171e1125fe7c2