Leak function in call bailout (#2466)
Summary:
Fixes #2464.

When we bail out of a function call because of a `FatalError` we did not leak the function we called. This means modified bindings were not appropriately leaked. See issue #2464 where the result is `undefined` instead of a reference to the leaked value.

This PR also adds a utility `__fatal` function for throwing a fatal error. Allowing us to more clearly express issues involving `FatalError`s. Used in the test.
Pull Request resolved: https://github.com/facebook/prepack/pull/2466

Differential Revision: D9482612

Pulled By: calebmer

fbshipit-source-id: 792d29818245ad9f631227d923efb56228df21e0