Fix stack overflow cleanup in pure function (#2433)
Summary:
Fixes #2432. I found this while trying to compile our internal React Native bundle.

Prepack was incorrectly cleaning up lexical scopes in the case of a stack overflow.

This bug is a bit difficult to reproduce. Follow these steps to reproduce:

1. Produce a stack overflow in code Prepack compiles. With a recursive loop for instance.
2. Wrap in `__evaluatePureFunction`. This should catch the stack overflow and emit a residual function call.
3. Add empty functions for `global.require` and `global.__d`. You need to do this so that we don’t crash [in tracer code trying to resolve these variables](https://github.com/facebook/prepack/blob/c18bd7ee97dc65b3e606ece5165854be698d2e95/src/methods/function.js#L82-L90) but we instead crash [in `PrepareForOrdinaryCall`](https://github.com/facebook/prepack/blob/c18bd7ee97dc65b3e606ece5165854be698d2e95/src/methods/function.js#L100).

I used a `try`/`catch` to fix this which is the obvious solution, but perhaps not the most efficient. Happy to write a slightly uglier solution for performance if we feel like it is necessary.
Pull Request resolved: https://github.com/facebook/prepack/pull/2433

Differential Revision: D9469599

Pulled By: calebmer

fbshipit-source-id: bd120c2e4bd450da3364c722550977d6566ed89a