Abstract empty values should fail RequireObjectCoercible (#2515)
Summary:
This fixes a bug it took me a while to track down. Unfortunately creating a repro was really difficult. I was observing the following invariant in `Reference` being triggered.

https://github.com/facebook/prepack/blob/fddfc6f0c52fa8f49fb0cb1208f34cefcb95dca3/src/environment.js#L1425-L1431

Digging in, the bad base was a deep conditional `AbstractValue` where all the leaf nodes were `EmptyValue` and `__bottomValue`. (Constructed from a number of `try`/`catch`/`throw`s.) This `AbstractValue` _should_ have simplified to just `EmptyValue` then `RequireObjectCoercible()` should have thrown a runtime `TypeError` after seeing the `EmptyValue`. This is what I observed in local tests when trying to reproduce this issue.

After looking at the error in the large React Native bundle I’m working in I discovered that the simplifier was throwing an error since the simplification count was exceeded. This is fine and normal. The issue was that `RequireObjectCoercible()` does not handle complex empty `AbstractValue`s well. If failing simplification is a normal part of Prepack operation we should make sure methods like `RequireObjectCoercible()` can handle complex `AbstractValue`s as well as simple ones.

This PR does that.
Pull Request resolved: https://github.com/facebook/prepack/pull/2515

Reviewed By: trueadm

Differential Revision: D9602146

Pulled By: calebmer

fbshipit-source-id: 359c74c646eeb619ba2e78b50eafe10b2e2fb133